separate annotation tests from `ClassFileImporterTest`
`ClassFileImporterTest` has meanwhile grown too much. It is hard to keep an overview of all the aspects covered and code analysis is sluggish. I have started to split it into several tests for different aspects, beginning with annotations.
